--- src/main/java/com/example/model/Admin.java
@@ -1,19 +1,0 @@
-package com.example.model;
-
-import lombok.Data;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-
-@Entity
-@Data
-public class Admin {
-
-    @Id
-    @GeneratedValue
-    private int id;
-
-    private String name;
-    private String password;
-}
