--- src/main/java/com/example/model/Student.java
@@ -1,42 +1,0 @@
-package com.example.model;
-
-import lombok.Data;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-
-@Entity
-public class Student {
-
-    @Id
-    @GeneratedValue
-    private int id;
-    //Teacher
-
-    @Column( name="name", length = 50)
-    private String name;
-    private String address;
-
-    public int getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getAddress() {
-        return address;
-    }
-
-    public void setAddress(String address) {
-        this.address = address;
-    }
-
-}
