--- src/main/java/com/example/model/Teacher.java
@@ -1,19 +1,0 @@
-package com.example.model;
-
-import lombok.Data;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-
-@Entity
-@Data
-public class Teacher {
-
-    @Id
-    @GeneratedValue
-    private int id;
-
-    private String name;
-    private String address;
-}
