--- src/main/java/com/example/repository/IAdminRepo.java
@@ -1,9 +1,0 @@
-package com.example.repository;
-
-import com.example.model.Admin;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface IAdminRepo extends JpaRepository<Admin, Integer> {
-}
