--- src/main/java/com/example/repository/IStudentRepo.java
@@ -1,9 +1,0 @@
-package com.example.repository;
-
-import com.example.model.Student;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface IStudentRepo extends JpaRepository <Student, Integer> {
-}
