--- src/main/java/com/example/repository/ITeacherRepo.java
@@ -1,9 +1,0 @@
-package com.example.repository;
-
-import com.example.model.Teacher;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ITeacherRepo extends JpaRepository<Teacher, Integer> {
-}
