+++ src/main/java/com/example/school/model/Student.java
@@ -1,0 +1,37 @@
+package com.example.school.model;
+
+import javax.persistence.*;
+
+@Entity
+public class Student {
+
+    @Id
+    @GeneratedValue(strategy= GenerationType.AUTO)
+    private int id;
+    //Teacher
+
+    @Column( name="name", length = 50)
+    private String name;
+    private String address;
+
+    public int getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getAddress() {
+        return address;
+    }
+
+    public void setAddress(String address) {
+        this.address = address;
+    }
+
+}
