+++ src/main/java/com/example/school/model/Teacher.java
@@ -1,0 +1,20 @@
+package com.example.school.model;
+
+import lombok.Data;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+
+@Entity
+@Data
+public class Teacher {
+
+    @Id
+    @GeneratedValue(strategy= GenerationType.AUTO)
+    private int id;
+
+    private String name;
+    private String address;
+}
