+++ src/main/java/com/example/school/repository/IAdminRepo.java
@@ -1,0 +1,9 @@
+package com.example.school.repository;
+
+import com.example.school.model.Admin;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface IAdminRepo extends JpaRepository<Admin, Integer> {
+}
