+++ src/main/java/com/example/school/repository/IStudentRepo.java
@@ -1,0 +1,9 @@
+package com.example.school.repository;
+
+import com.example.school.model.Student;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface IStudentRepo extends JpaRepository <Student, Integer> {
+}
