+++ src/main/java/com/example/school/repository/ITeacherRepo.java
@@ -1,0 +1,9 @@
+package com.example.school.repository;
+
+import com.example.school.model.Teacher;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ITeacherRepo extends JpaRepository<Teacher, Integer> {
+}
