+++ src/main/java/com/example/school/service/IStudentService.java
@@ -1,0 +1,8 @@
+package com.example.school.service;
+
+import com.example.school.model.Student;
+
+public interface IStudentService {
+    boolean save(Student student) throws Exception;
+    boolean findAll();
+}
