+++ src/main/resources/application.properties
@@ -1,4 +1,4 @@
-spring.datasource.url=jdbc:mysql://localhost:3306/school?useSSL=false&serverTimezone=America/Mexico_City
+spring.datasource.url=jdbc:mysql://localhost:3306/school
 spring.datasource.username=root
 spring.datasource.password=admin
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
